item

+++ cogs/leveling/leveling.py
@@ -69,12 +69,16 @@ class Leveling(commands.Cog):
     async def on_guild_remove(self, guild):
         for member in guild.members:
             if not member.bot:
-                await levelling.delete_one({"guild": guild.id, "user": member.id})
+                result = await levelling.find_one({"guild": guild.id, "user": member.id})
+                if result is not None:
+                    await levelling.delete_one({"guild": guild.id, "user": member.id})
 
     @commands.Cog.listener()
     async def on_member_remove(self, member):
         if not member.bot:
-            await levelling.delete_one({"guild": member.guild.id, "user": member.id})
+            result = await levelling.find_one({"guild": member.guild.id, "user": member.id})
+            if result is not None:
+                await levelling.delete_one({"guild": member.guild.id, "user": member.id})
 
 
 def setup(bot):
